+++ tests/MEDS_Transforms/test_add_time_derived_measurements.py
@@ -3,11 +3,17 @@
 Set the bash env variable `DO_USE_LOCAL_SCRIPTS=1` to use the local py files, rather than the installed
 scripts.
 """
+import rootutils
+
+root = rootutils.setup_root(__file__, dotenv=True, pythonpath=True, cwd=True)
 
 from meds import subject_id_field
 
-from .transform_tester_base import ADD_TIME_DERIVED_MEASUREMENTS_SCRIPT, single_stage_transform_tester
-from .utils import parse_meds_csvs
+from tests.MEDS_Transforms.transform_tester_base import (
+    ADD_TIME_DERIVED_MEASUREMENTS_SCRIPT,
+    single_stage_transform_tester,
+)
+from tests.utils import parse_meds_csvs
 
 AGE_CALCULATION_STR = """
 See `add_time_derived_measurements.py` for the source of the constant value.
