+++ tests/MEDS_Transforms/test_fit_vocabulary_indices.py
@@ -4,8 +4,12 @@ Set the bash env variable `DO_USE_LOCAL_SCRIPTS=1` to use the local py files, ra
 scripts.
 """
 
+import rootutils
 
-from .transform_tester_base import (
+root = rootutils.setup_root(__file__, dotenv=True, pythonpath=True, cwd=True)
+
+
+from tests.MEDS_Transforms.transform_tester_base import (
     FIT_VOCABULARY_INDICES_SCRIPT,
     parse_code_metadata_csv,
     single_stage_transform_tester,
