+++ tests/MEDS_Transforms/test_occlude_outliers.py
@@ -4,11 +4,14 @@ Set the bash env variable `DO_USE_LOCAL_SCRIPTS=1` to use the local py files, ra
 scripts.
 """
 
+import rootutils
+
+rootutils.setup_root(__file__, dotenv=True, pythonpath=True, cwd=True)
 
 import polars as pl
 
-from .transform_tester_base import OCCLUDE_OUTLIERS_SCRIPT, single_stage_transform_tester
-from .utils import MEDS_PL_SCHEMA, parse_meds_csvs
+from tests.MEDS_Transforms.transform_tester_base import OCCLUDE_OUTLIERS_SCRIPT, single_stage_transform_tester
+from tests.utils import MEDS_PL_SCHEMA, parse_meds_csvs
 
 # This is the code metadata
 # MEDS_CODE_METADATA_CSV = """
